Include error position information in ParseException (goes some way to addressing issue #73)
I changed constructors on ParseException - the parameterless one is
standard for Exception classes but isn't very useful for a library that
has excellent messages and didn't appear to be used anywhere, so I
removed it; the other two constructors now receive an "errorPosition"
constructor since everywhere in the library that raises this exception
knows where the problem occurred.

I guess that, strictly speaking, this could be considered a breaking
change since any external code that raised its own ParseExceptions would
need changing.. I wonder how likely that is considered to be? An
alternative would be to restore the constructors that don't take an
"errorPosition" argument and to leave set "ErrorPosition" to
"ErrorPosition.Empty" if those constructors are called, though this
would require anyone interrogating the property to have to check the
"HasValue" property and conclude that this information isn't always
available (when I think that it should be!)